Refactor FavoritesController; restrict access to user favorites and remove edit/show functionality

+++ src/Controller/Frontend/FavoritesController.php
@@ -10,10 +10,12 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/favorites')]
 final class FavoritesController extends AbstractController{
     #[Route(name: 'app_favorites_index', methods: ['GET'])]
+    #[IsGranted('ROLE_USER')]
     public function index(FavoritesRepository $favoritesRepository): Response
     {
         return $this->render('favorites/index.html.twig', [
@@ -22,9 +24,11 @@ final class FavoritesController extends AbstractController{
     }
 
     #[Route('/new', name: 'app_favorites_new', methods: ['GET', 'POST'])]
+    #[IsGranted('ROLE_USER')]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $favorite = new Favorites();
+        $favorite->setUser($this->getUser());
         $form = $this->createForm(FavoritesType::class, $favorite);
         $form->handleRequest($request);
 
@@ -41,35 +45,14 @@ final class FavoritesController extends AbstractController{
         ]);
     }
 
-    #[Route('/{id}', name: 'app_favorites_show', methods: ['GET'])]
-    public function show(Favorites $favorite): Response
-    {
-        return $this->render('favorites/show.html.twig', [
-            'favorite' => $favorite,
-        ]);
-    }
-
-    #[Route('/{id}/edit', name: 'app_favorites_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Favorites $favorite, EntityManagerInterface $entityManager): Response
-    {
-        $form = $this->createForm(FavoritesType::class, $favorite);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->flush();
-
-            return $this->redirectToRoute('app_favorites_index', [], Response::HTTP_SEE_OTHER);
-        }
-
-        return $this->render('favorites/edit.html.twig', [
-            'favorite' => $favorite,
-            'form' => $form,
-        ]);
-    }
-
     #[Route('/{id}', name: 'app_favorites_delete', methods: ['POST'])]
+    #[IsGranted('ROLE_USER')]
     public function delete(Request $request, Favorites $favorite, EntityManagerInterface $entityManager): Response
     {
+        if ($favorite->getUser() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('You can only delete your own favorite.');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$favorite->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($favorite);
             $entityManager->flush();
